empanadas: Normalize file writing to use with blocks
We found some explicit `close()` calls when writing files in a `with`
block, which is not necessary.  Files are closed when exiting the block,
either by normal execution or by exception.

We also convert a few instances of file writes *not* using `with`
blocks, to use the appropriate `with` syntax.